requirements.txt
streamlit==1.31.0
openai==1.12.0
langgraph==0.0.30
langchain==0.1.10
langchain-openai==0.0.8
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
plotly==5.18.0
asyncio==3.4.3
httpx==0.26.0
